+++ interfaces/report-type.ts
@@ -2,6 +2,4 @@ export type Report = {
   reportName: string;
   reportDate: string;
   reportFormat: string;
-  isGenerated: boolean;
-  data: object[];
 }